requirements.txt
# Django
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Task Queue & Caching
celery==5.3.4
redis==5.0.1

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Authentication & Security
PyJWT==2.8.0
python-jose==3.3.0

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Email & Markdown
markdown==3.5.2
markdown2==2.4.9

# Utilities
python-slugify==8.0.1
pytz==2023.3

# Development
black==23.12.1
flake8==6.1.0
isort==5.13.2
pylint==3.0.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
